Assert zip adapters return values
Unit tests did not properly tests that both adapters had the exact
same behavior for returned values. By using common code for testing
we can assert that the behavior is now in fact identical.

Also phpdocs were moved to the interface itself instead of each
adapters.
